requirements.txt
dbus-python==1.2.16
pycairo==1.20.1
PyGObject==3.40.1
